Refactored to share logic between extensions
