+++ directus-cms/extensions-src/publishable/.gitignore
@@ -1,0 +1,3 @@
+.DS_Store
+node_modules
+dist
