+++ directus-cms/extensions-src/publishable/src/presentation-publishable.vue
@@ -1,12 +1,11 @@
 <script>
 import { useApi } from '@directus/extensions-sdk';
-import { isPublishable } from './isPublishable.js';
+import { isPublishable } from './../../../extensions/shared/isPublishable.js';
 
 export default {
   props: {
     collection: null,
     primaryKey: null,
-    debug: true,
   },
   data() {
     return {
@@ -19,16 +18,10 @@ export default {
       this.loadFields(this.collection),
     ])
     .then(([itemResponse, fieldResponse]) => {
-      // Access data from the first and second responses
       const itemData = itemResponse.data.data;
       const fieldData = fieldResponse.data.data;
 
-      // Proceed with further processing using both data
-      console.log('itemData data:', itemData);
-      console.log('fieldData data:', fieldData);
-
       let testResult = isPublishable(itemData, fieldData);
-      console.log('isPublishable', testResult);
 
       this.publishable = testResult;
     })
@@ -54,19 +47,20 @@ export default {
 </script>
 
 <template>
-  <p>Publishable: {{ publishable }}</p>
-  <p>collection: {{ collection }}</p>
-  <p>primaryKey: {{ primaryKey }}</p>
   <div v-if='publishable === true'>
+    <v-icon name="check_circle" left />
     Can be published.
   </div>
   <div v-else-if='publishable === false'>
+    <v-icon name="unpublished" left />
     Can not be published.
   </div>
   <div v-else-if='publishable === null'>
+    <v-progress-circular indeterminate />
     Publishable state not yet determined...
   </div>
   <div v-else>
+    <v-icon name="warning" left />
     Unknown publishable state. This should not happen and indicates an error.
   </div>
 </template>
